+++ docs/reference-1.6.0.html
@@ -24880,7 +24880,7 @@ var Leaflet = L.noConflict();
 
 	</div>
 
-	<script src="../docs/docs/js/docs.js"></script>
-	<script src="../docs/docs/js/reference.js"></script>
+	<script src="docs/tsc/docs.js"></script>
+	<script src="docs/tsc/reference.js"></script>
 
 </body></html>
