+++ spec/.eslintrc
@@ -11,6 +11,6 @@
     "sinon": false,
     "happen": false,
     "Hand": false,
-    "touchEventType": false /* defined in SpecHelper.js */
+    "touchEventType": false /* defined in SpecHelper.tsc */
   }
 }
