+++ src/geo/LatLng.ts
@@ -10,7 +10,7 @@ import {toLatLngBounds} from './LatLngBounds';
  * @example
  *
  * ```
- * var latlng = L.latLng(50.5, 30.5);
+ * const latlng = L.latLng(50.5, 30.5);
  * ```
  *
  * All Leaflet methods that accept LatLng objects also accept them in a simple Array form and simple object form (unless noted otherwise), so these lines are equivalent:
@@ -55,7 +55,7 @@ LatLng.prototype = {
 
 		obj = toLatLng(obj);
 
-		var margin = Math.max(
+		const margin = Math.max(
 		        Math.abs(this.lat - obj.lat),
 		        Math.abs(this.lng - obj.lng));
 
@@ -85,7 +85,7 @@ LatLng.prototype = {
 	// @method toBounds(sizeInMeters: Number): LatLngBounds
 	// Returns a new `LatLngBounds` object in which each boundary is `sizeInMeters/2` meters apart from the `LatLng`.
 	toBounds: function (sizeInMeters) {
-		var latAccuracy = 180 * sizeInMeters / 40075017,
+		const latAccuracy = 180 * sizeInMeters / 40075017,
 		    lngAccuracy = latAccuracy / Math.cos((Math.PI / 180) * this.lat);
 
 		return toLatLngBounds(
