+++ src/geo/LatLngBounds.ts
@@ -8,7 +8,7 @@ import {LatLng, toLatLng} from './LatLng';
  *
  * @example
  *
- * ```js
+ * ```tsc
  * var corner1 = L.latLng(40.712, -74.227),
  * corner2 = L.latLng(40.774, -74.125),
  * bounds = L.latLngBounds(corner1, corner2);
@@ -16,7 +16,7 @@ import {LatLng, toLatLng} from './LatLng';
  *
  * All Leaflet methods that accept LatLngBounds objects also accept them in a simple Array form (unless noted otherwise), so the bounds example above can be passed like this:
  *
- * ```js
+ * ```tsc
  * map.fitBounds([
  * 	[40.712, -74.227],
  * 	[40.774, -74.125]
