+++ src/layer/tile/TileLayer.WMS.ts
@@ -12,7 +12,7 @@ import {toBounds} from '../../geometry/Bounds';
  *
  * @example
  *
- * ```js
+ * ```tsc
  * const nexrad = L.tileLayer.wms("http://mesonet.agron.iastate.edu/cgi-bin/wms/nexrad/n0r.cgi", {
  * 	layers: 'nexrad-n0r-900913',
  * 	format: 'image/png',
