+++ src/layer/tile/TileLayer.ts
@@ -13,7 +13,7 @@ import * as DomUtil from '../../dom/DomUtil';
  *
  * @example
  *
- * ```js
+ * ```tsc
  * L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png?{foo}', {foo: 'bar', attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'}).addTo(map);
  * ```
  *
