+++ src/server.ts
@@ -1,20 +1,22 @@
 // 'use strict'; // eslint https://stackoverflow.com/questions/32791507/node-js-and-eslint-disagree-on-use-strict
 
-// require('dist/leaflet-src.ts');
+require('./src/Leaflet.ts');
 
 import express from 'express';
-export require('iconv').Iconv;
+// export require('iconv').Iconv;
 // const express = require('express');
 
-// Constants
-declare const PORT = 8080;
-declare const HOST = '0.0.0.0';
+function server(){
+  // Constants
+  const PORT = 8080;
+  const HOST = '0.0.0.0';
 
 // App
-const app = express();
-app.get('/', (req, res) => {
-  res.send('Hello World');
-});
+  const app = express();
+  app.get('/', (req, res) => {
+    res.send('Hello World');
+  });
 
-app.listen(PORT, HOST);
-console.log(`Running on http://${HOST}:${PORT}`);
+  app.listen(PORT, HOST);
+  console.log(`Running on http://${HOST}:${PORT}`);
+}
